Introduce conversion parameter
This is an example of Parameterize Method (Refactoring, p283). I think.
It's about the closest Refactoring pattern I see to what I'm doing here.

Before, the conversion methods just *knew* what Roman symbols to use.
This made them impossible to use for the neo-roman feature. By making a
constant that handles Roman conversion, and then changing the methods to
work with that conversion constant, we can easily add new conversions..