Consolidate methods with Parameterize Method (?)
When we tried to implement exoroman numerals we had to add new methods
to handle its need to display numbers higher than 3999. This refactoring
will make implementing that behavior easier while following OCP.

I put a ? after Parameterize Method because it's not really a perfect
match for what we did. Its description

> several methods do similar things but with different values contained
> in the method body. Create one method that uses a parameter for the
> different values

Is pretty close to what we did here, but not exact. We had several
methods that had the exact same logic (a case statement that switched on
the number value when divided by 1/10/100/1000). We have consolidated
all that logic in one method.

This new method takes advantage of the fact that the `conversion`
elements have powers-of-ten every other key/value pair. This is
certainly hacky and will most certainly bite us.

That this method is gross is not exactly a problem at this point. What
we're trying to do is be able to implement exoroman as easily as
possible.