Extract classes RomanConverter and DeciRomans
Two changes in this commit.

First, extracting DeciRomans. This was done because of the problems we
ran into implementing duodeciromans. The current conversion method was
hard-coded to use multiples of 10 and 5. Extracting out those numbers to
a class we can use as needed.

Second, extracting RomanConverter. As we added the concept of a
dictionary and a conversion 'family', keeping all of this logic in
Fixnum just got ridiculous.